Add BridgeBase contract

+++ contracts/IToken.sol
@@ -3,7 +3,7 @@
 pragma solidity ^0.8.19;
 
 interface IToken {
-    function mint(address to, uint amount) external;
+    function mint(address to, uint256 amount) external;
 
-    function burn(address owner, uint amount) external;
+    function burn(address owner, uint256 amount) external;
 }
